+++ docs/RELEASE-NOTES.rst
@@ -1,14 +1,17 @@
 Release Notes for Container Ingress Services for Kubernetes & OpenShift
 =======================================================================
 
-2.8.0
+Next Release
 -------------
 Added Functionality
 ```````````````````
 
 **What’s new:**
     * CRD:
-        * Persistence Profile CR support for VirtualServer, TransportServer and Policy CRs. `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
+        * Persistence Profile support for VirtualServer, TransportServer and Policy CRs. `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
+        * Added support for host in TransportServer and IngressLink CR. `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
+        * Added support for multiple health monitors in EDNS resource
+    * NodePortLocal(NPL) antrea cni feature support added to Ingress and CRD Resources.
 
 Enhancements
 ````````````